modified mega header
