+++ apps/nextjs-website/src/components/MegaHeader.tsx
@@ -3,7 +3,8 @@ import React from 'react';
 import MegaHeaderRC from '@react-components/components/MegaHeader/MegaHeader';
 import { MegaHeaderProps } from '@react-components/types/MegaHeader/MegaHeader.types';
 import { MegaHeaderSection } from '@/lib/fetch/types/PageSection';
-import { defaultMenuItems, defaultSocialMediaLinks } from '../../stories/MegaHeader/megaheaderCommons';
+import { defaultMenuItems, defaultSocialMediaLinks } from '../../stories/MegaHeader/light.stories';
+
 
 
 const makeMegaHeaderProps = ({
